feat: create SearchBar Component, separate logic
