+++ src/App.tsx
@@ -3,6 +3,7 @@ import classNames from "classnames/bind";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import CampingList from "./components/CampingList/CampingList";
 import Map from "./components/Map/Map";
+import SearchBar from "./components/SearchBar/SearchBar";
 
 const cx = classNames.bind(styles);
 
@@ -12,6 +13,7 @@ function App() {
   return (
     <QueryClientProvider client={queryClient}>
       <div className={cx("container")}>
+        <SearchBar />
         <Map />
         <CampingList />
       </div>
