+++ src/components/SearchBar/SearchBar.module.scss
@@ -1,0 +1,8 @@
+@import "../../styles/utils";
+
+.container {
+  width: 100%;
+  height: rem(52);
+  display: flex;
+  align-items: center;
+}
