+++ src/components/UI/Button/Button.module.scss
@@ -17,7 +17,7 @@
   }
 
   &:disabled {
-    color: var(--gray3);
+    color: var(--white);
     background-color: var(--gray3);
     border-color: var(--gray3);
     pointer-events: none;
